+++ test/main/Test/Format/ParseTime.hs
@@ -16,6 +16,7 @@ import Data.Proxy
 import Data.Time
 import Data.Time.Calendar.OrdinalDate
 import Data.Time.Calendar.WeekDate
+import Data.Time.Calendar.MonthCount
 import Test.Arbitrary ()
 import Test.QuickCheck.Property
 import Test.Tasty
@@ -573,6 +574,7 @@ readShowTests =
     nameTest
         "read_show"
         [ nameTest "Day" (prop_read_show :: Day -> Result)
+        , nameTest "Month" (prop_read_show :: Month -> Result)
         , nameTest "TimeOfDay" (prop_read_show :: TimeOfDay -> Result)
         , nameTest "LocalTime" (prop_read_show :: LocalTime -> Result)
         , nameTest "TimeZone" (prop_read_show :: TimeZone -> Result)