+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## [1.11]
 - new calendrical type synonyms and abstract constructors
+- new absolute Month type, with appropriate functions
 - new functions for working with week-based years
 - new parseTimeMultipleM function for a list of (format, input) pairs
 - add instance Ord DayOfWeek